feat: improved file structure + event loop singleton structure
