--- asyncio/asyncio_utils.py
@@ -1,18 +1,0 @@
-from typing import Awaitable, Coroutine, TypeVar
-
-from .future import Future
-from .event_loop import EventLoop
-
-
-T = TypeVar("T")
-
-
-def ensure_future(awaitable: Awaitable[T], *, loop: EventLoop) -> Future[T]:
-    # https://peps.python.org/pep-0492/
-    if isinstance(awaitable, Future):
-        return awaitable
-    if isinstance(awaitable, Coroutine):
-        return loop.create_task(awaitable)
-    raise TypeError(
-        f"object provided is not a Future or Corotuine. Type: {type(awaitable).__name__}"
-    )
